REMOVE T-MOBILE THEME ENGINE (2/3)
As it stands right now, this is only functional on 2-3 devices and for some reason, is not on about 4-5 devices.

I've never seen anything like this before but it is obvious that there are some conflicts that need to be fixed.

With that said, I want it out of DU for now since is obviously not stable and until we can fix the conflicts.

Change-Id: I6c35ed026e3e1378ceab2fdab155c923622562a7