manifest: README: Try to clear up ambiguity over Gerrit push command
Some people get confused about what they are supposed to change in this
command, because HEAD is capitalized (I know I was). Use lowercase and
keep variables in order in the example.

Also, provide the link to Gerrit projects in case there is any
confusion over what it should be.

Change-Id: Ib985ceaa41438df430be338df76e380e6ea8d2a9
Signed-off-by: Nathan Chancellor <nathan@dirtyunicorns.com>