Re-track the media hw/qcom repo
Why the back/forth? Because we initially thought that we didn't have the
problem we have on the marlin with the printf format but turns out that
we do. We tried tracking it down and see what's causing this issue but
were unsuccessful. So tracking this baby again because is just easier.

Change-Id: I28499bed51f377f165090c35f3e6760509739fba